refactor: simplify immediately invoked function detection
