+++ test/rules/sort-variable-declarations.test.ts
@@ -860,152 +860,6 @@ describe('sort-variable-declarations', () => {
     testDependencyDetection(false)
 
     describe('experimental detection specific', () => {
-      it('detects dependencies inside immediately invoked class methods', async () => {
-        await valid({
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          code: dedent`
-            let b = 1,
-                a = new (class { m() { return b } })().m();
-          `,
-        })
-
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsDependencyOrder',
-              data: { nodeDependentOnRight: 'a', right: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let b = 1,
-                a = new (class { m() { return b } })().m();
-          `,
-          code: dedent`
-            let a = new (class { m() { return b } })().m(),
-                b = 1;
-          `,
-        })
-      })
-
-      it('detects dependencies inside immediately invoked object methods', async () => {
-        await valid({
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          code: dedent`
-            let b = 1,
-                a = ({ m() { return b } }).m();
-          `,
-        })
-      })
-
-      it('detects dependencies inside immediately used getters', async () => {
-        await valid({
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          code: dedent`
-            let b = 1,
-                a = ({ get x() { return b } }).x;
-          `,
-        })
-      })
-
-      it('detects dependencies inside immediately invoked class constructors', async () => {
-        await valid({
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          code: dedent`
-            let b = 1,
-                a = new (class { constructor() { this.x = b } })().x;
-          `,
-        })
-      })
-
-      it('detects dependencies inside immediately invoked static class methods', async () => {
-        await valid({
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          code: dedent`
-            let b = 1,
-                a = (class { static m() { return b } }).m();
-          `,
-        })
-      })
-
-      it('ignores deferred references to object and class members', async () => {
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = ({ m() { return b } }).m,
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = ({ m() { return b } }).m;
-          `,
-        })
-
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = class { m() { return b } },
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = class { m() { return b } };
-          `,
-        })
-      })
-
       it('ignores dependencies inside class declarations in immediately invoked functions', async () => {
         await invalid({
           errors: [
@@ -1027,104 +881,6 @@ describe('sort-variable-declarations', () => {
           code: dedent`
             let b = 1,
                 a = (function() { class D { m() { return b } } return 1 })();
-          `,
-        })
-      })
-
-      it('ignores members of object literals that are not accessed', async () => {
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = { m() { return b } },
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = { m() { return b } };
-          `,
-        })
-      })
-
-      it('ignores class expression constructors that are not instantiated', async () => {
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = (class { constructor() { this.x = b } }).x,
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = (class { constructor() { this.x = b } }).x;
-          `,
-        })
-      })
-
-      it('ignores functions used as computed keys', async () => {
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = { [() => b]: 0 },
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = { [() => b]: 0 };
-          `,
-        })
-
-        await invalid({
-          errors: [
-            {
-              messageId: 'unexpectedVariableDeclarationsOrder',
-              data: { right: 'a', left: 'b' },
-            },
-          ],
-          options: [
-            {
-              ...options,
-              useExperimentalDependencyDetection: true,
-            },
-          ],
-          output: dedent`
-            let a = class { [() => b]() {} },
-                b = 1;
-          `,
-          code: dedent`
-            let b = 1,
-                a = class { [() => b]() {} };
           `,
         })
       })
